+++ src/commands/zsets/zmpop.ts
@@ -233,7 +233,7 @@ async function blockingZsetMultiPop(
 export const zmpopCommand = defineCommand({
   name: 'zmpop',
   since: { redis: '7.0.0', valkey: '7.2.0' },
-  schema: t.custom<ZsetMultiPopArgs>((input, index, ctx) => ({
+  schema: t.custom<ZsetMultiPopArgs>({ min: 3 }, (input, index, ctx) => ({
     value: parseZsetMultiPopArgs(input, index, ctx, { blocking: false }),
     nextIndex: input.length,
   })),
@@ -247,10 +247,13 @@ export const zmpopCommand = defineCommand({
 export const bzmpopCommand = defineCommand({
   name: 'bzmpop',
   since: { redis: '7.0.0', valkey: '7.2.0' },
-  schema: t.custom<BlockingZsetMultiPopArgs>((input, index, ctx) => ({
-    value: parseZsetMultiPopArgs(input, index, ctx, { blocking: true }),
-    nextIndex: input.length,
-  })),
+  schema: t.custom<BlockingZsetMultiPopArgs>(
+    { min: 4 },
+    (input, index, ctx) => ({
+      value: parseZsetMultiPopArgs(input, index, ctx, { blocking: true }),
+      nextIndex: input.length,
+    }),
+  ),
   flags: ['write', 'noscript'],
   keys: args => args.keys,
   execute: (args, ctx) => {